update talent pools
